backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
scikit-learn==1.3.0
exifread==3.0.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
gunicorn==21.2.0
sqlalchemy==2.0.19
flask-sqlalchemy==3.0.5
pyjwt==2.8.0
python-magic==0.4.27
requests==2.31.0
